test: split PDF equation prose separation coverage
